requirements.txt
biopython>=1.81
pandas>=2.0.0
requests>=2.31.0
numpy>=1.24.0
matplotlib=3.10.0 # ""
seaborn=0.13.2
tqdm=4.67.1
scipy=1.15.1